1. Updated compatibility to arduino-esp8266 3.0.2 release 2. Added blink gpio mode for test

+++ src/interface/WiFiClientInterface.cpp
@@ -172,7 +172,7 @@ uint16_t  WiFiClientInterface::localPort(){
 /**
  * availableForWrite
  */
-size_t WiFiClientInterface::availableForWrite(){
+int WiFiClientInterface::availableForWrite(){
   return this->m_wifi_client.availableForWrite();
 }
